+++ src/neuroconv/tools/roiextractors/imagingextractordatachunkiterator.py
@@ -41,7 +41,7 @@ class ImagingExtractorDataChunkIterator(GenericDataChunkIterator):
             The upper bound on size in megabytes (MB) of the internal chunk for the HDF5 dataset.
             The chunk_shape will be set implicitly by this argument.
             Cannot be set if `chunk_shape` is also specified.
-            The default is 1MB, as recommended by the HDF5 group. For more details, see
+            The default is 10MB. For more details, see
             https://support.hdfgroup.org/HDF5/doc/TechNotes/TechNote-HDF5-ImprovingIOPerformanceCompressedDatasets.pdf
         chunk_shape : tuple, optional
             Manual specification of the internal chunk shape for the HDF5 dataset.
@@ -62,12 +62,12 @@ class ImagingExtractorDataChunkIterator(GenericDataChunkIterator):
             assert chunk_mb * 1e6 <= buffer_gb * 1e9, "chunk_mb must be less than or equal to buffer_gb!"
 
         if chunk_mb is None and chunk_shape is None:
-            chunk_mb = 1.0
+            chunk_mb = 10.0
 
         self._maxshape = self._get_maxshape()
         self._dtype = self._get_dtype()
         if chunk_shape is None:
-            chunk_shape = super()._get_default_chunk_shape(chunk_mb=chunk_mb)
+            chunk_shape = self._get_default_chunk_shape(chunk_mb=chunk_mb)
 
         if buffer_gb is None and buffer_shape is None:
             buffer_gb = 1.0
@@ -81,6 +81,18 @@ class ImagingExtractorDataChunkIterator(GenericDataChunkIterator):
             display_progress=display_progress,
             progress_bar_options=progress_bar_options,
         )
+
+    def _get_default_chunk_shape(self, chunk_mb: float) -> tuple:
+        """Select the chunk_shape less than the threshold of chunk_mb while keeping the original image size."""
+        assert chunk_mb > 0, f"chunk_mb ({chunk_mb}) must be greater than zero!"
+
+        num_frames = self._maxshape[0]
+        image_shape = self._maxshape[1:]
+        frame_size_bytes = math.prod(image_shape) * self._dtype.itemsize
+        chunk_size_bytes = chunk_mb * math.prod(image_shape)
+        num_frames_per_chunk = int(chunk_size_bytes // frame_size_bytes)
+        chunk_shape = (max(min(num_frames_per_chunk, num_frames), 1), *image_shape)
+        return chunk_shape
 
     def _get_scaled_buffer_shape(self, buffer_gb: float, chunk_shape: tuple) -> tuple:
         """Select the buffer_shape less than the threshold of buffer_gb that is also a multiple of the chunk_shape."""